+++ examples/assets_basic_machine_learning/assets_basic_machine_learning/__init__.py
@@ -1,0 +1,9 @@
+from dagster import Definitions, load_assets_from_modules
+
+from . import assets
+
+all_assets = load_assets_from_modules([assets])
+
+defs = Definitions(
+    assets=all_assets,
+)
