+++ examples/assets_basic_machine_learning/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+name = assets_basic_machine_learning
