--- examples/assets_machine_learning/machine_learning_example/basic_machine_learning_example/__init__.py
@@ -1,9 +1,0 @@
-from dagster import Definitions, load_assets_from_modules
-
-from . import assets
-
-all_assets = load_assets_from_modules([assets])
-
-defs = Definitions(
-    assets=all_assets,
-)
