--- examples/assets_machine_learning/machine_learning_example/machine_learning_example_tests/__init__.py
@@ -1,1 +1,0 @@
-
