--- examples/assets_machine_learning/machine_learning_example/machine_learning_example_tests/test_assets.py
@@ -1,1 +1,0 @@
-
