--- examples/assets_machine_learning/machine_learning_example/machine_learning_with_mlflow/__init__.py
@@ -1,9 +1,0 @@
-from dagster import Definitions, load_assets_from_modules
-
-from . import mlflow_assets
-
-all_assets = load_assets_from_modules([mlflow_assets])
-
-defs = Definitions(
-    assets=all_assets,
-)
