--- examples/assets_machine_learning/machine_learning_example/pyproject.toml
@@ -1,6 +1,0 @@
-[build-system]
-requires = ["setuptools"]
-build-backend = "setuptools.build_meta"
-
-[tool.dagster]
-module_name = "machine_learning_example"
