--- examples/assets_machine_learning/machine_learning_example/setup.cfg
@@ -1,2 +1,0 @@
-[metadata]
-name = machine_learning_example
