--- examples/assets_machine_learning/machine_learning_example/workspace.yaml
@@ -1,5 +1,0 @@
-load_from:
-  - python_package:
-      package_name: basic_machine_learning_example
-  - python_package:
-      package_name: machine_learning_with_mlflow
