+++ examples/assets_mlflow_machine_learning/pyproject.toml
@@ -1,0 +1,6 @@
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
+[tool.dagster]
+module_name = "assets_mlflow_machine_learning"
