+++ examples/assets_mlflow_machine_learning/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+name = assets_mlflow_machine_learning
